services/extract/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AWS Services
boto3==1.37.3
botocore==1.37.3
aiobotocore==2.22.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
greenlet==3.0.3

# Message Queue
aio-pika==9.5.5

# LLM and AI
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.6
langchain-core==0.1.23

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78

# HTTP client and file handling
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
tenacity==8.2.3

# JSON and data processing
orjson==3.9.12
jsonschema==4.20.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-docker==2.0.1
moto==5.1.5

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 